Improved Spawn System
To ensure that the number of monsters spawned is below the limit set in each GameLevel, I had added a variable called currentMonsters which didn't accurately measure how many monsters there were in the level.

I changed this so that the number of monsters were measured using an ArrayList called monsters.